+++ nudebomb/walk.py
@@ -1,11 +1,11 @@
 """Walk directory trees and strip mkvs."""
 import os
-
 from copy import deepcopy
 from pathlib import Path
 
 from termcolor import cprint
-from treestamps import Treestamps
+from treestamps import Grovestamps, GrovestampsConfig
+from treestamps.tree import Treestamps
 
 from nudebomb.config import TIMESTAMPS_CONFIG_KEYS
 from nudebomb.langfiles import LangFiles
@@ -20,18 +20,14 @@ class Walk:
         """Initialize."""
         self._config = config
         self._langfiles = LangFiles(config)
-        self._timestamps: dict[Path, Treestamps] = {}
 
     def _is_path_ignored(self, path: Path) -> bool:
         """Return if path should be ignored."""
-        for ignore_glob in self._config.ignore:
-            if path.match(ignore_glob):
-                return True
-        return False
+        return any(path.match(ignore_glob) for ignore_glob in self._config.ignore)
 
     def strip_path(self, top_path, path):
         """Strip a single mkv file."""
-        if not path.suffix == ".mkv":
+        if path.suffix != ".mkv":
             return
 
         mtime = None
@@ -47,24 +43,24 @@ class Walk:
                 cprint(".", "cyan", end="")
             return
 
-        dirpath = Treestamps.dirpath(path)
+        dir_path = Treestamps.get_dir(path)
         config = deepcopy(self._config)
-        config.languages = self._langfiles.get_langs(top_path, dirpath)
+        config.languages = self._langfiles.get_langs(top_path, dir_path)
         mkv_obj = MKVFile(config, path)
         mkv_obj.remove_tracks()
 
         if self._config.timestamps:
-            self._timestamps[top_path].set(path)
+            self._timestamps[top_path.resolve()].set(path)
 
-    def walk_dir(self, top_path, dir):
+    def walk_dir(self, top_path, dir_path):
         """Walk a directory."""
         if not self._config.recurse:
             return
 
         filenames = []
 
-        for filename in sorted(os.listdir(dir)):
-            entry_path = dir / filename
+        for filename in sorted(os.listdir(dir_path)):
+            entry_path = dir_path / filename
             if entry_path.is_dir():
                 self.walk_file(top_path, entry_path)
             else:
@@ -74,8 +70,8 @@ class Walk:
             self.walk_file(top_path, path)
 
         if self._config.timestamps:
-            timestamps = self._timestamps[top_path]
-            timestamps.set(dir, compact=True)
+            timestamps = self._timestamps[top_path.resolve()]
+            timestamps.set(dir_path, compact=True)
 
     def walk_file(self, top_path, path):
         """Walk a file."""
@@ -99,41 +95,39 @@ class Walk:
     def print_info(self):
         """Print intentions before we begin."""
         langs = ", ".join(sorted(self._config.languages))
-        if self._config.sub_languages:
-            audio = "audio "
-        else:
-            audio = ""
-        print(f"Stripping {audio}languages except {langs}.")
+        audio = "audio " if self._config.sub_languages else ""
+        cprint(f"Stripping {audio}languages except {langs}.")
         if self._config.sub_languages:
             sub_langs = ", ".join(sorted(self._config.sub_languages))
-            print(f"Stripping subtitle languages except {sub_langs}.")
+            cprint(f"Stripping subtitle languages except {sub_langs}.")
 
-        print("Searching for MKV files to process", end="")
+        cprint("Searching for MKV files to process", end="")
         if self._config.verbose:
-            print(":")
+            cprint(":")
 
     def run(self):
         """Run the stripper against all configured paths."""
         self.print_info()
 
         if self._config.timestamps:
-            self._timestamps = Treestamps.map_factory(
-                self._config.paths,
+            copse_config = GrovestampsConfig(
                 PROGRAM_NAME,
-                self._config.verbose,
-                self._config.symlinks,
-                self._config.ignore,
-                self._config,
-                TIMESTAMPS_CONFIG_KEYS,
+                paths=self._config.paths,
+                verbose=self._config.verbose,
+                symlinks=self._config.symlinks,
+                ignore=self._config.ignore,
+                check_config=self._config.timestamps_check_config,
+                program_config=self._config,
+                program_config_keys=TIMESTAMPS_CONFIG_KEYS,
             )
+            self._timestamps = Grovestamps(copse_config)
+
         for path_str in self._config.paths:
             path = Path(path_str)
-            top_path = Treestamps.dirpath(path)
+            top_path = Treestamps.get_dir(path)
             self.walk_file(top_path, path)
         if not self._config.verbose:
-            print("done.")
+            cprint("done.")
 
         if self._config.timestamps:
-            for top_path, timestamps in self._timestamps.items():
-                print(f"Saving timestamps for {top_path}")
-                timestamps.dump()
+            self._timestamps.dump()