+++ run.py
@@ -2,6 +2,5 @@
 """Development runner."""
 from nudebomb.cli import main
 
-
 if __name__ == "__main__":
     main()