+++ tests/util.py
@@ -1,9 +1,7 @@
 """Common test utilities."""
 import json
 import subprocess
-
 from pathlib import Path
-
 
 TEST_FN = "test5.mkv"
 SRC_DIR = Path("tests/test_files")
@@ -13,8 +11,9 @@ __all__ = ()
 
 
 def mkv_tracks(path):
+    """Get tracks from mkv."""
     cmd = ("mkvmerge", "-J", str(path))
-    proc = subprocess.run(cmd, check=True, capture_output=True, text=True)
+    proc = subprocess.run(cmd, check=True, capture_output=True, text=True)  # noqa S603
     data = json.loads(proc.stdout)
     return data.get("tracks")